+++ dev/benchmarks/test_apps/stocks/lib/i18n/stock_strings.dart
@@ -1,7 +1,3 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
@@ -9,6 +5,10 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 // ignore: unused_import
 import 'package:intl/intl.dart' as intl;
+
+import 'stock_strings_en.dart';
+import 'stock_strings_en_US.dart';
+import 'stock_strings_es.dart';
 
 // ignore_for_file: unnecessary_brace_in_string_interps
 
@@ -65,10 +65,10 @@ import 'package:intl/intl.dart' as intl;
 /// be consistent with the languages listed in the StockStrings.supportedLocales
 /// property.
 abstract class StockStrings {
-  StockStrings(String locale) : assert(locale != null), _localeName = intl.Intl.canonicalizedLocale(locale.toString());
+  StockStrings(String locale) : assert(locale != null), localeName = intl.Intl.canonicalizedLocale(locale.toString());
 
   // ignore: unused_field
-  final String _localeName;
+  final String localeName;
 
   static StockStrings of(BuildContext context) {
     return Localizations.of<StockStrings>(context, StockStrings);
@@ -123,48 +123,6 @@ class _StockStringsDelegate extends LocalizationsDelegate<StockStrings> {
 
   @override
   bool shouldReload(_StockStringsDelegate old) => false;
-}
-
-/// The translations for English (`en`).
-class StockStringsEn extends StockStrings {
-  StockStringsEn([String locale = 'en']) : super(locale);
-
-  @override
-  String get title => 'Stocks';
-
-  @override
-  String get market => 'MARKET';
-
-  @override
-  String get portfolio => 'PORTFOLIO';
-}
-
-/// The translations for English, as used in the United States (`en_US`).
-class StockStringsEnUs extends StockStringsEn {
-  StockStringsEnUs([String locale = 'en_US']) : super(locale);
-
-  @override
-  String get title => 'Stocks';
-
-  @override
-  String get market => 'MARKET';
-
-  @override
-  String get portfolio => 'PORTFOLIO';
-}
-
-/// The translations for Spanish Castilian (`es`).
-class StockStringsEs extends StockStrings {
-  StockStringsEs([String locale = 'es']) : super(locale);
-
-  @override
-  String get title => 'Acciones';
-
-  @override
-  String get market => 'MERCADO';
-
-  @override
-  String get portfolio => 'CARTERA';
 }
 
 StockStrings _lookupStockStrings(Locale locale) {
