+++ dev/benchmarks/test_apps/stocks/lib/i18n/stock_strings_es.dart
@@ -1,0 +1,19 @@
+// ignore: unused_import
+import 'package:intl/intl.dart' as intl;
+import 'stock_strings.dart';
+
+// ignore_for_file: unnecessary_brace_in_string_interps
+
+/// The translations for Spanish Castilian (`es`).
+class StockStringsEs extends StockStrings {
+  StockStringsEs([String locale = 'es']) : super(locale);
+
+  @override
+  String get title => 'Acciones';
+
+  @override
+  String get market => 'MERCADO';
+
+  @override
+  String get portfolio => 'CARTERA';
+}
